moved the Digest- and Cipher_IO_Stream usage pattern test cases to UsagePatternTestExtra.java
